removing alerts
